% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compfun.R
\name{compfun}
\alias{compfun}
\title{Component Functions Estimates}
\usage{
compfun(A, y, a, s, t)
}
\arguments{
\item{A}{Aggregated dataset matrix.}

\item{y}{Weights of the component functions matrix.}

\item{a}{The weight of the point mass function at zero in the prior.}

\item{s}{The standard deviation of the random error.}

\item{t}{The scale parameter of the logistic distribution in the prior.}
}
\value{
The component functions estimates matrix.
}
\description{
Estimates component functions from aggregated curves by applying wavelet shrinkage under logistic prior.
}
\examples{
A = matrix(rnorm(80),16,5)
y1 = runif(5)
y = t(matrix(c(y1,1-y1),5,2))
compfun(A,y,0.9,1,5)
}
