% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbias.R
\name{logbias}
\alias{logbias}
\title{Bias of the logistic shrinkage rule.}
\usage{
logbias(theta, a, s, t)
}
\arguments{
\item{theta}{The wavelet coefficients vector.}

\item{a}{The weight of the point mass function at zero in the prior.}

\item{s}{The standard deviation of the random error.}

\item{t}{The scale parameter of the logistic distribution in the prior.}
}
\value{
The bias vector.
}
\description{
Performs the bias of the wavelet shrinkage rule under logistic prior.
}
\examples{
logbias(c(0,0.2,0.5),0.9,1,5)
}
