% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbrisk.R
\name{logbrisk}
\alias{logbrisk}
\title{Bayesian risk of the logistic shrinkage rule}
\usage{
logbrisk(a, s, t)
}
\arguments{
\item{a}{The weight of the point mass function at zero in the prior.}

\item{s}{The standard deviation of the random error.}

\item{t}{The scale parameter of the logistic distribution in the prior.}
}
\value{
The bayesian risk value.
}
\description{
Performs the bayesian risk of the wavelet shrinkage rule under logistic prior.
}
\examples{
logbrisk(0.9,1,5)

}
