% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrisk.R
\name{logrisk}
\alias{logrisk}
\title{Classical risk of the logistic shrinkage rule}
\usage{
logrisk(theta, a, s, t)
}
\arguments{
\item{theta}{The wavelet coefficients vector.}

\item{a}{The weight of the point mass function at zero in the prior.}

\item{s}{The standard deviation of the random error.}

\item{t}{The scale parameter of the logistic distribution in the prior.}
}
\value{
The classical risks vector.
}
\description{
Performs the classical risks of the wavelet shrinkage rule under logistic prior.
}
\examples{
logrisk(c(0,0.2,0.5),0.9,1,5)

}
