% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrule.R
\name{logrule}
\alias{logrule}
\title{Shrinkage Rule under Logistic Prior}
\usage{
logrule(d, a, s, t)
}
\arguments{
\item{d}{The empirical wavelet coefficients vector.}

\item{a}{The weight of the point mass function at zero in the prior.}

\item{s}{The standard deviation of the random error.}

\item{t}{The scale parameter of the logistic distribution in the prior.}
}
\value{
The shrunk wavelet coefficients vector.
}
\description{
Performs wavelet shrinkage under logistic prior.
}
\examples{
logrule(c(2,1,2.3,0.5,0.1,-0.2,4.8,0.2),0.9,1,5)

}
