% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logvar.R
\name{logvar}
\alias{logvar}
\title{Variance of the logistic shrinkage rule.}
\usage{
logvar(theta, a, s, t)
}
\arguments{
\item{theta}{The wavelet coefficients vector.}

\item{a}{The weight of the point mass function at zero in the prior.}

\item{s}{The standard deviation of the random error.}

\item{t}{The scale parameter of the logistic distribution in the prior.}
}
\value{
The variances vector.
}
\description{
Performs the variance of the wavelet shrinkage rule under logistic prior.
}
\examples{
logvar(c(0,0.2,0.5),0.9,1,5)

}
